requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.0

streamlit==1.39.0

python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3

selenium==4.38.0
webdriver-manager==4.0.2

numpy==2.1.2

python-docx==1.2.0
lxml==6.0.2

pymupdf==1.24.10

openai==1.55.0
google-generativeai==0.7.2
tiktoken==0.7.0
httpx==0.27.2

typing-extensions==4.15.0
matplotlib==3.9.2